pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "media-consolidation-tool"
version = "2.0.0"
description = "Enhanced media consolidation and review tool with checkpoint support"
authors = [{name = "Media Tool Team"}]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "Pillow>=8.0.0",
    "imagehash>=4.0.0",
    "tqdm>=4.50.0",
]

[project.scripts]
media-tool = "media_tool.main:main"
